# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "titile": "Plans Wizard's API docs",
    "description": "API 명세서",
    "version": "1.0.0",
    "title": "REST API"
  },
  "host": "localhost:3000",
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "http",
      "scheme": "bearer",
      "in": "header",
      "bearerFormat": "JWT"
    }
  },
  "paths": {
    "/auth/join": {
      "post": {
        "tags": [
          "authRouter"
        ],
        "summary": "로컬 회원가입 API",
        "description": "사용자가 로컬 회원가입할 때 사용하는 엔드포인트",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "회원가입할 유저 정보",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "user_name"
                },
                "email": {
                  "type": "string",
                  "example": "user_email"
                },
                "password": {
                  "type": "string",
                  "example": "user_password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "정상적으로 회원가입 처리됨"
          },
          "400": {
            "description": "잘못된 요청 양식"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "authRouter"
        ],
        "summary": "로컬 로그인 API",
        "description": "사용자가 로컬 로그인할 때 사용하는 엔드포인트",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "로그인할 유저 정보",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "user_email"
                },
                "password": {
                  "type": "string",
                  "example": "user_password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "정상적으로 로그인 처리됨"
          },
          "400": {
            "description": "잘못된 요청 양식"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "tags": [
          "authRouter"
        ],
        "summary": "로그아웃 API",
        "description": "사용자가 로그아웃할 때 사용하는 엔드포인트",
        "responses": {
          "200": {
            "description": "정상적으로 로그인 처리됨"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/auth/google": {
      "get": {
        "tags": [
          "authRouter"
        ],
        "summary": "구글 로그인 API",
        "description": "사용자가 구글 로그인할 때 사용하는 엔드포인트",
        "responses": {
          "200": {
            "description": "정상적으로 로그인 처리됨"
          }
        }
      }
    },
    "/auth/google/callback": {
      "get": {
        "tags": [
          "authRouter"
        ],
        "summary": "구글 로그인 리다이렉트 API",
        "description": "클라이언트가 구글 인증 서버에서 인증 코드를 받기 위한 엔드포인트, 클라이언트로부터 요청이 다이렉트로 들어오지 않는 주소",
        "responses": {
          "200": {
            "description": "정상적으로 로그아웃됨"
          }
        }
      }
    },
    "/group/create": {
      "post": {
        "tags": [
          "groupRouter"
        ],
        "summary": "그룹 생성 API",
        "description": "사용자가 그룹을 생성할 때 사용하는 엔드포인트",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "생성할 그룹 정보",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "group_name"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "정상적으로 그룹이 생성됨"
          },
          "400": {
            "description": "잘못된 요청 양식"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        }
      }
    },
    "/group/:id/profile": {
      "get": {
        "tags": [
          "groupRouter"
        ],
        "summary": "그룹 정보 조회 API",
        "description": "그룹의 정보를 조회할 때 사용하는 엔드포인트",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "정상적으로 그룹 정보가 조회됨"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        }
      }
    },
    "/group/:id/users": {
      "get": {
        "tags": [
          "groupRouter"
        ],
        "summary": "그룹 참여자 정보 조회 API",
        "description": "그룹 참여자 정보를 조회할 때 사용하는 엔드포인트",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "정상적으로 그룹 참여자 정보가 조회됨"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        }
      }
    },
    "/preference": {
      "get": {
        "tags": [
          "preferenceRouter"
        ],
        "summary": "선호도 조회 API",
        "description": "사용자가 자신의 선호도를 조회할 때 사용하는 엔드포인트",
        "responses": {
          "200": {
            "description": "정상적으로 선호도 조회됨"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        }
      },
      "post": {
        "tags": [
          "preferenceRouter"
        ],
        "summary": "선호도 수정 API",
        "description": "사용자가 자신의 선호도를 수정할 때 사용하는 엔드포인트",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "수정할 선호도 정보",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "day_preference": {
                  "type": "object",
                  "properties": {
                    "Mon": {
                      "type": "number",
                      "example": 3
                    },
                    "Tue": {
                      "type": "number",
                      "example": 3
                    },
                    "Wed": {
                      "type": "number",
                      "example": 3
                    },
                    "Thu": {
                      "type": "number",
                      "example": 3
                    },
                    "Fri": {
                      "type": "number",
                      "example": 3
                    },
                    "Sat": {
                      "type": "number",
                      "example": 3
                    },
                    "Sun": {
                      "type": "number",
                      "example": 3
                    }
                  }
                },
                "time_preference": {
                  "type": "object",
                  "properties": {
                    "am": {
                      "type": "number",
                      "example": 3
                    },
                    "pm": {
                      "type": "number",
                      "example": 3
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "정상적으로 선호도 수정됨"
          },
          "400": {
            "description": "잘못된 요청 양식"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "서버 내부 오류"
          }
        }
      }
    },
    "/test/session": {
      "get": {
        "tags": [
          "testRouter"
        ],
        "summary": "세션 실험용 API",
        "description": "로그인 세션 확인용 엔드포인트",
        "responses": {
          "200": {
            "description": "요청 성공 응답"
          }
        }
      }
    }
  }
}